+++ packages/af-utils-react-virtual-headless/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/react-virtual-headless
 
+## 0.0.20
+
+### Patch Changes
+
+-   improved scrollTo
+
 ## 0.0.19
 
 ### Patch Changes